+++ src/main/java/com/ongil/backend/domain/review/service/prompter/ReviewPrompter.java
@@ -1,0 +1,8 @@
+package com.ongil.backend.domain.review.service.prompter;
+
+import com.ongil.backend.domain.review.dto.request.AiReviewGenerateRequest;
+
+public interface ReviewPrompter {
+	String getSystemPrompt();
+	String buildUserMessage(AiReviewGenerateRequest request);
+}
